Revert change to hash_bigint
I made a mistake fixing frozenset hashing. It turns out that the oddly
small modulus is correct whenever a PyInt needs to be hashed, but we
still want to avoid re-hashing the hash value of a frozenset. The
solution is to leave `hash_bigint` as it was and simply avoid calling it
in `slot::hash_wrapper`.